# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"APIResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"errors": {},
					"reason": {
						"type": "string"
					},
					"status": {
						"type": "string",
						"enum": [
							"success",
							"failed"
						]
					}
				},
				"required": [
					"status"
				],
				"type": "object"
			},
			"TextPayload": {
				"properties": {
					"body": {
						"type": "string"
					},
					"type": {
						"type": "string",
						"enum": [
							"text"
						],
						"nullable": false
					}
				},
				"required": [
					"body",
					"type"
				],
				"type": "object"
			},
			"MediaPayload": {
				"properties": {
					"text": {
						"type": "string"
					},
					"url": {
						"type": "string"
					},
					"type": {
						"type": "string",
						"enum": [
							"image",
							"video",
							"audio",
							"document"
						]
					}
				},
				"required": [
					"url",
					"type"
				],
				"type": "object"
			},
			"ParameterPayload": {
				"properties": {
					"payload": {},
					"type": {
						"type": "string",
						"enum": [
							"parameter"
						],
						"nullable": false
					}
				},
				"type": "object"
			},
			"MessageRequest": {
				"properties": {
					"payload": {
						"items": {
							"anyOf": [
								{
									"$ref": "#/components/schemas/TextPayload"
								},
								{
									"$ref": "#/components/schemas/MediaPayload"
								},
								{
									"$ref": "#/components/schemas/ParameterPayload"
								}
							]
						},
						"type": "array"
					},
					"provider": {
						"type": "string"
					},
					"type": {
						"type": "string",
						"enum": [
							"text",
							"media"
						]
					},
					"to": {
						"type": "string"
					}
				},
				"required": [
					"payload",
					"provider",
					"type",
					"to"
				],
				"type": "object"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "openapi",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/generic/{username}": {
			"post": {
				"operationId": "Hello",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse"
								}
							}
						}
					}
				},
				"tags": [
					"generic"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "username",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/generic": {
			"get": {
				"operationId": "Foo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse"
								}
							}
						}
					}
				},
				"tags": [
					"generic"
				],
				"security": [],
				"parameters": []
			}
		},
		"/messaging/sendMessage": {
			"post": {
				"operationId": "SendMessage",
				"responses": {
					"202": {
						"description": "Accepted",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse"
								}
							}
						}
					}
				},
				"tags": [
					"messaging"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MessageRequest"
							}
						}
					}
				}
			}
		},
		"/user/login/{username}": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse"
								}
							}
						}
					}
				},
				"tags": [
					"user"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "username",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/user/{username}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse"
								}
							}
						}
					}
				},
				"tags": [
					"user"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "username",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}